requirements.txt
alembic==1.0.0
argcomplete==1.9.3
asn1crypto==0.24.0
astroid==2.2.4
awscli==1.16.170
backcall==0.1.0
bcrypt==3.1.4
bleach==3.0.2
boto3==1.9.112
botocore==1.12.112
CacheControl==0.12.5
cachetools==3.1.0
certifi==2018.11.29
cffi==1.12.2
cfn-flip==1.2.0
chardet==3.0.4
Click==7.0
colorama==0.4.1
coverage==4.5.1
cryptography==2.6.1
cycler==0.10.0
decorator==4.3.0
defusedxml==0.5.0
Django==2.2.4
docutils==0.14
durationpy==0.5
entrypoints==0.2.3
firebase-admin==2.16.0
Flask==1.0.2
Flask-API==1.0
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.7
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
future==0.16.0
google-api-core==1.8.0
google-auth==1.6.3
google-cloud-core==0.29.1
google-cloud-firestore==0.31.0
google-cloud-storage==1.14.0
google-resumable-media==0.3.2
googleapis-common-protos==1.5.8
gripcontrol==3.3.0
grpcio==1.19.0
gunicorn==19.9.0
hjson==3.0.1
idna==2.8
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.12
itsdangerous==1.1.0
jedi==0.13.1
Jinja2==2.10
jmespath==0.9.4
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
kappa==0.6.0
kiwisolver==1.0.1
lambda-packages==0.20.0
lazy-object-proxy==1.3.1
lxml==4.3.2
Mako==1.0.7
MarkupSafe==1.1.1
matplotlib==3.0.2
mccabe==0.6.1
mistune==0.8.4
msgpack==0.6.1
nbconvert==5.4.0
nbformat==4.4.0
neo4j-driver==1.6.3
neomodel==3.3.1
neotime==1.7.4
notebook==5.7.0
numpy==1.15.4
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pickleshare==0.7.5
placebo==0.9.0
prometheus-client==0.4.2
prompt-toolkit==2.0.7
protobuf==3.7.0
psycopg2==2.7.5
pubcontrol==2.4.2
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycodestyle==2.5.0
pycparser==2.19
pydocstyle==3.0.0
pyflakes==2.1.1
Pygments==2.2.0
PyJWT==1.6.4
pylama==7.7.1
pylint==2.1.1
pyOpenSSL==19.0.0
pyparsing==2.3.0
python-dateutil==2.8.0
python-dotenv==0.10.1
python-editor==1.0.3
python-slugify==1.2.4
pytz==2018.9
pywinpty==0.5.4
PyYAML==3.13
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.21.0
requests-toolbelt==0.9.1
rsa==4.0
s3transfer==0.2.0
scikit-learn==0.20.0
scipy==1.1.0
Send2Trash==1.5.0
simplejson==3.16.0
six==1.12.0
snowballstemmer==1.2.1
SQLAlchemy==1.2.12
SQLAlchemy-Utils==0.34.1
sqlparse==0.3.0
terminado==0.8.1
testpath==0.4.2
toml==0.10.0
tornado==5.1.1
tqdm==4.19.1
traitlets==4.3.2
troposphere==2.4.6
typed-ast==1.3.1
Unidecode==1.0.23
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.11.1
wsgi-request-logger==0.4.6
zappa==0.48.2
